requirements.txt
chromadb
langchain
google-generativeai
langchain-google-genai
tiktoken
streamlit
numpy
python-dotenv
requests
pandas
streamlit